[lib][hwkey] Fix the build warning in hwkey.c

+++ lib/hwkey/hwkey.c
@@ -109,7 +109,7 @@ static long send_req(hwkey_session_t session, struct hwkey_msg *msg, uint8_t *re
 
 	if (inf.len > sizeof(*msg) + *rsp_buf_len) {
 		TLOGE("%s: insufficient output buffer size (%zu > %zu)\n",
-		      __func__, inf.len - sizeof(*msg), *rsp_buf_len);
+		      __func__, inf.len - sizeof(*msg), (size_t)*rsp_buf_len);
 		rc = ERR_TOO_BIG;
 		goto err_get_fail;
 	}
